+++ src/pipeline/preprocess/query/enrichBQLQuery.ts
@@ -153,6 +153,9 @@ export const enrichBQLQuery: PipelineOperation<BaseResponse> = async (req) => {
 				$idNotIncluded: idNotIncluded,
 				$filterByUnique: checkFilterByUnique(field.$filter, currentSchema),
 				$filterProcessed: true,
+				$sort: field.$sort,
+				$offset: field.$offset,
+				$limit: field.$limit,
 			};
 		});
 	};
@@ -209,6 +212,9 @@ export const enrichBQLQuery: PipelineOperation<BaseResponse> = async (req) => {
 				$filterByUnique: checkFilterByUnique(field.$filter, currentSchema),
 				$playedBy: playedBy,
 				$filterProcessed: true,
+				$sort: field.$sort,
+				$offset: field.$offset,
+				$limit: field.$limit,
 			};
 		});
 	};